Moved menu into it's own component
