+++ components/eventCard/EventItem.tsx
@@ -1,67 +1,28 @@
-import React, { useState } from "react"
-import {
-  Alert,
-  Image,
-  StyleSheet,
-  Text,
-  TouchableWithoutFeedback,
-  View
-} from "react-native"
+import React from "react"
+import { Image, StyleSheet, Text, View } from "react-native"
 import { Event } from "@lib/events/Event"
 import { Divider } from "react-native-elements"
 import { Shadow } from "react-native-shadow-2"
 import tinycolor from "tinycolor2"
-import {
-  Menu,
-  MenuOptions,
-  MenuOption,
-  MenuTrigger
-} from "react-native-popup-menu"
 import { placemarkToFormattedAddress } from "@lib/location"
-import { TouchableOpacity } from "react-native-gesture-handler"
 import { compactFormatMiles } from "@lib/DistanceFormatting"
-import { MaterialCommunityIcon, MaterialIcon } from "../common/Icons"
+import { MaterialCommunityIcon } from "../common/Icons"
+import MenuDropdown from "./MenuDropdown"
 
 interface Props {
   event: Event
 }
 
 const EventItem = ({ event }: Props) => {
-  const [isMenuOpen, setMenuOpen] = useState(false)
   const numAttendees = 1 // Eventually will come from event object
   const distance = 0.5 // Eventually will be calculated from the location given in the event object
   const lightEventColor = tinycolor(event.color).lighten(25).toString()
-  let dismiss: () => void
-
-  const alertDelete = () => {
-    dismissMenu()
-    Alert.alert("Delete this event?", undefined, [
-      { text: "Delete", style: "destructive", onPress: dismiss },
-      { text: "Cancel", style: "cancel", onPress: dismiss }
-    ])
-  }
-
-  // Will probably need to use a modal instead to have user select/type in reason for report
-  const alertReport = () => {
-    dismissMenu()
-    Alert.alert("Report this event?", undefined, [
-      { text: "Report", style: "destructive", onPress: dismiss },
-      { text: "Cancel", style: "cancel" }
-    ])
-  }
-
-  const editForm = () => {
-    console.log("call event form")
-  }
 
   // Show event details screen
   const onPressEvent = () => {
     console.log("test")
   }
 
-  const dismissMenu = () => {
-    setMenuOpen(false)
-  }
   return (
     <Shadow
       distance={16}
@@ -82,54 +43,7 @@ const EventItem = ({ event }: Props) => {
 
           {/* Dropdown menu when more button is pressed */}
           <View style={styles.moreButtonStyle}>
-            <Menu
-              opened={isMenuOpen}
-              onBackdropPress={() => setMenuOpen(false)}
-            >
-              <MenuTrigger
-                testID="more options"
-                onPress={() => setMenuOpen(true)}
-                customStyles={{
-                  TriggerTouchableComponent: TouchableOpacity
-                }}
-                // Increase area where menu is triggered
-                style={{
-                  paddingLeft: 16,
-                  paddingVertical: 6
-                }}
-              >
-                <MaterialIcon name="more-horiz" />
-              </MenuTrigger>
-              {/* Need to use Text in order for tests to find the element
-                Also need to move the onSelect functions to child in order to test functionality
-                */}
-              {event.writtenByYou
-                ? (
-                  <MenuOptions customStyles={menuOptionsStyles}>
-                    <MenuOption>
-                      <Text style={{ color: "red" }} onPress={alertDelete}>
-                      Delete
-                      </Text>
-                    </MenuOption>
-                    <MenuOption>
-                      <Text onPress={editForm}>Edit</Text>
-                    </MenuOption>
-                  </MenuOptions>
-                )
-                : (
-                  <MenuOptions
-                    customStyles={{
-                      optionsContainer: {
-                        width: "30%"
-                      }
-                    }}
-                  >
-                    <MenuOption>
-                      <Text onPress={alertReport}>Report</Text>
-                    </MenuOption>
-                  </MenuOptions>
-                )}
-            </Menu>
+            <MenuDropdown isEventHost={event.writtenByYou} />
           </View>
         </View>
 
@@ -274,19 +188,7 @@ const styles = StyleSheet.create({
     flex: 1,
     opacity: 0.4,
     alignItems: "flex-end"
-  },
-  optionsText: {
-    textAlignVertical: "center",
-    fontSize: 16
-  },
-  optionsContainer: {
-    width: "30%"
   }
 })
-
-const menuOptionsStyles = {
-  optionsContainer: styles.optionsContainer,
-  optionsText: styles.optionsText
-}
 
 export default EventItem
